% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{as_SpatRaster}
\alias{as_SpatRaster}
\alias{as_SpatRaster.satres}
\title{As \code{terra} \code{SpatRaster} class}
\usage{
as_SpatRaster(sr, res)

\method{as_SpatRaster}{satres}(sr, res = NULL)
}
\arguments{
\item{sr}{A \code{satres} object.}

\item{res}{A string, spatial resolution.}
}
\value{
A vector of strings.
}
\description{
Returns as an object of class \code{SpatRaster} from package \code{terra} the
multi-band raster of the indicated spatial resolution.
}
\examples{
# tsr <- sr |>
#        as_SpatRaster("r10m")

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite functions: 
\code{\link{get_spatial_resolutions}()},
\code{\link{satres}()},
\code{\link{save_by_resolution}()}
}
\concept{satellite functions}
