% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{check_bands}
\alias{check_bands}
\title{Check band}
\usage{
check_bands(sr, res = NULL, bands = NULL)
}
\arguments{
\item{sr}{A \code{satres} object.}

\item{res}{A string, spatial resolution.}

\item{bands}{A string, band names.}
}
\value{
boolean.
}
\description{
Check if the indicated band is one of those available.
}
\keyword{internal}
