% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{find_name_to_files}
\alias{find_name_to_files}
\title{find name to files}
\usage{
find_name_to_files(files, patterns)
}
\arguments{
\item{files}{A string vector.}

\item{patterns}{A string vector of values with names.}
}
\value{
A string vector.
}
\description{
Finds the name associated to a file name in a vector of named patterns.
}
\keyword{internal}
