% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{get_spatial_resolutions}
\alias{get_spatial_resolutions}
\alias{get_spatial_resolutions.satres}
\title{Get spatial resolutions}
\usage{
get_spatial_resolutions(sr)

\method{get_spatial_resolutions}{satres}(sr)
}
\arguments{
\item{sr}{A \code{satres} object.}
}
\value{
A vector of strings.
}
\description{
Returns the spatial resolutions of the multi-band raster that make up the object.
}
\examples{
# res <- sr |>
#        get_spatial_resolutions()

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite functions: 
\code{\link{as_SpatRaster}()},
\code{\link{satres}()},
\code{\link{save_by_resolution}()}
}
\concept{satellite functions}
