% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{get_spectral_resolution}
\alias{get_spectral_resolution}
\alias{get_spectral_resolution.satres}
\title{Get spectral resolutions}
\usage{
get_spectral_resolution(sr, res)

\method{get_spectral_resolution}{satres}(sr, res = NULL)
}
\arguments{
\item{sr}{A \code{satres} object.}

\item{res}{A string, spatial resolution.}
}
\value{
A vector of strings.
}
\description{
Returns the spectral resolutions of the multi-band raster that make up the
object or, what is the same, the band names.
}
\details{
We can indicate the name of a certain spatial resolution to obtain only its
band names.
}
\examples{

esa <- system.file("extdata", "esa", package = "satres")
sr <- satres(dir = esa)

r <- sr |>
     get_spectral_resolution()

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite definition: 
\code{\link{get_all_names}()},
\code{\link{get_band_names}()},
\code{\link{get_spatial_resolution}()},
\code{\link{satres}()}
}
\concept{satellite definition}
