% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sat_aggregate}
\alias{sat_aggregate}
\title{Aggregate rasters in a folder}
\usage{
sat_aggregate(dir, out_dir, factor = 100)
}
\arguments{
\item{dir}{A string or string vector.}

\item{out_dir}{A string, output folder.}

\item{factor}{A integer.}
}
\value{
A string vector.
}
\description{
Select the files that correspond to the band selection. Each one is assigned
the band identifier as a name.
}
\keyword{internal}
