% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat_band}
\alias{sat_band}
\title{Final part of the name and extension of the satellite band files}
\format{
A vector.
}
\usage{
sat_band
}
\description{
Additionally, the name of each element is the band identifier. There will be
repeated names (bands with the same identifier) due to having different
spatial resolution.
}
\seealso{
Other satellite data: 
\code{\link{sat_rest}}
}
\concept{satellite data}
\keyword{datasets}
