% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sat_dimensions}
\alias{sat_dimensions}
\title{Get dimensions}
\usage{
sat_dimensions(sr)
}
\arguments{
\item{sr}{A \code{satres} object.}
}
\value{
A string vector.
}
\description{
Get all dimensions in vector form to check them.
}
\keyword{internal}
