% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sat_minmax}
\alias{sat_minmax}
\title{Get minmax}
\usage{
sat_minmax(sr)
}
\arguments{
\item{sr}{A \code{satres} object.}
}
\value{
A string vector.
}
\description{
Get all minmax values in vector form to check them.
}
\keyword{internal}
