% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sat_names}
\alias{sat_names}
\title{Get names}
\usage{
sat_names(sr)
}
\arguments{
\item{sr}{A \code{satres} object.}
}
\value{
A string vector.
}
\description{
Get all names in vector form to check them.
}
\keyword{internal}
