% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_res.R
\name{sat_res}
\alias{sat_res}
\title{Unzip compressed files in tar or zip format}
\usage{
sat_res(dir, out_dir = NULL, only_bands = TRUE)
}
\arguments{
\item{dir}{A string or string vector.}

\item{out_dir}{A string or string vector, output folder.}

\item{only_bands}{A boolean, include only satellite bands.}
}
\value{
A vector of strings, name of the processed files.
}
\description{
Given a vector of compressed file names or the name of a folder containing
compressed files, unzip the files to the given output folder. If no output
folder is indicated, it is considered the same folder where the input files
are.
}
\details{
We can indicate whether to include the file name (without the extension) as a
folder in the output folder.
}
\examples{

# sat_untarzip("dat/usgs")

# sat_untarzip("dat/esa")

# file <- c("dat/usgs/LC08_L1TP_200034_20230924_20230924_02_RT.tar",
#           "dat/esa/S2A_MSIL2A_20230905T105621_N0509_R094_T30SVF_20230905T170700.zip")
# sat_untarzip(file)

}
\concept{satellite functions}
