% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat_rest}
\alias{sat_rest}
\title{Final part of the name and extension of satellite rasters that are not bands}
\format{
A vector.
}
\usage{
sat_rest
}
\description{
Additionally, the name of each element is the raster identifier. There will be
repeated names (rasters with the same identifier) due to having different
spatial resolution.
}
\seealso{
Other satellite data: 
\code{\link{sat_band}}
}
\concept{satellite data}
\keyword{datasets}
