% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sat_rest_msk}
\alias{sat_rest_msk}
\title{Mask of name of satellite rasters that are not bands}
\format{
A vector.
}
\usage{
sat_rest_msk
}
\description{
Raster name patterns to treat and not consider bands.
}
\seealso{
Other satellite data: 
\code{\link{sat_band}},
\code{\link{sat_rest}}
}
\concept{satellite data}
\keyword{datasets}
