% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{save_by_resolution}
\alias{save_by_resolution}
\alias{save_by_resolution.satres}
\title{Get the page information of a pivot table}
\usage{
save_by_resolution(sr, out_dir)

\method{save_by_resolution}{satres}(sr, out_dir = NULL)
}
\arguments{
\item{sr}{A \code{satres} object.}

\item{out_dir}{A string, output folder.}
}
\value{
A \code{satres} object.
}
\description{
Get the page information associated with the pivot table represented by the
object.
}
\details{
Each pivot table implements a report. The pivot table page represents the
context of that report. It is useful when we work with several pivot tables
with the same structure: for example, the page can allow us to differentiate
their origin, date or author. This information is often included in the file
name, sheet name, or cells attached to the pivot table.
}
\examples{
# sr <- sr |>
#       save_by_resolution()

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite functions: 
\code{\link{satres}()}
}
\concept{satellite functions}
