% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{select_band_files}
\alias{select_band_files}
\title{Select band files}
\usage{
select_band_files(files)
}
\arguments{
\item{files}{A string vector.}
}
\value{
A list of string vectors.
}
\description{
Select the files that correspond to the band selection. Each one is assigned
the band identifier as a name.
}
\keyword{internal}
