% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{transform_to_multiband}
\alias{transform_to_multiband}
\title{Transform to multi band}
\usage{
transform_to_multiband(bands)
}
\arguments{
\item{bands}{A list of raster bands.}
}
\value{
A raster band.
}
\description{
Transforms a list of raster bands in a raster band.
}
\keyword{internal}
