% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{satres}
\alias{satres}
\title{\code{satres} S3 class}
\usage{
satres(dir, out_dir = NULL, only_spectral = TRUE)
}
\arguments{
\item{dir}{A string or string vector, folder names.}

\item{out_dir}{A string, output folder.}

\item{only_spectral}{A boolean, include only spectral bands.}
}
\value{
A \code{satres} object.
}
\description{
Creates a \code{satres} object from a set of raster files.
}
\details{
Given a folder name or a vector of folder names, containing satellite band
raster files, creates an object containing all rasters grouped according to
their spatial resolution.

If there are several rasters of the same area (tiles), it previously merges
them to form a single raster of the total area.

A working folder where the virtual rasters are created can be indicated as a
parameter. Additionally, we indicate whether we wish to process only the
spectral band files (B1 to B12) or all available files.
}
\examples{

esa <- system.file("extdata", "esa", package = "satres")

sr <- satres(dir = esa)

sr <- satres(dir = esa,
             out_dir = tempdir(),
             only_spectral = FALSE)

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite definition: 
\code{\link{get_band_names}()},
\code{\link{get_spatial_resolution}()},
\code{\link{get_spectral_band_names}()}
}
\concept{satellite definition}
