% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{save_by_resolution}
\alias{save_by_resolution}
\alias{save_by_resolution.satres}
\title{Save multi-band rasters according to their spatial resolution}
\usage{
save_by_resolution(sr, out_dir)

\method{save_by_resolution}{satres}(sr, out_dir = NULL)
}
\arguments{
\item{sr}{A \code{satres} object.}

\item{out_dir}{A string, output folder.}
}
\value{
A \code{satres} object.
}
\description{
Saves multi-band raster files of the object according to its spatial
resolution. The file names correspond to the resolution of each one.
}
\details{
They are stored in the folder that is indicated or, if none is indicated, in
the folder that was used to create the object.
}
\examples{
# sr <- sr |>
#       save_by_resolution()

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite functions: 
\code{\link{as_SpatRaster}()},
\code{\link{get_spatial_resolutions}()},
\code{\link{satres}()}
}
\concept{satellite functions}
