% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satres.R
\name{transform_to_multiband}
\alias{transform_to_multiband}
\title{Transforms a list of raster bands in a raster multi-band}
\usage{
transform_to_multiband(bands)
}
\arguments{
\item{bands}{A list of raster bands.}
}
\value{
A raster band.
}
\description{
To generate a multi-band raster, the definition of the vector must be
executed in a single instruction, which is why it must be done through
meta programming.
}
\keyword{internal}
